+++ tools/endpoint_server_example.cpp
@@ -113,7 +113,7 @@ int main()
       new dlms::cosem::CosemRegisterObject(
         dlms::cosem::CosemLogicalName(1, 0, 1, 8, 0, 255),
         EncodeLongUnsigned(2300u),
-        dlms::cosem::CosemByteBuffer(),
+        dlms::cosem::types::ScalerUnit(),
         dlms::cosem::AttributeAccessMode::ReadOnly)));
 
   dlms::endpoint::ServerEndpoint endpoint(channel, logicalDevice);